+++ src/core/ObservableQuery.ts
@@ -38,6 +38,7 @@ export type ApolloCurrentResult<T> = {
   loading: boolean;
   networkStatus: NetworkStatus;
   error?: ApolloError;
+  partial?: boolean;
 };
 
 export interface FetchMoreOptions {
@@ -115,6 +116,12 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
     });
   }
 
+  /**
+   * Return the result of the query from the local cache as well as some fetching status
+   * `loading` and `networkStatus` allow to know if a request is in flight
+   * `partial` lets you know if the result from the local cache is complete or partial
+   * @return {result: Object, loading: boolean, networkStatus: number, partial: boolean}
+   */
   public currentResult(): ApolloCurrentResult<T> {
     const { data, partial } = this.queryManager.getCurrentQueryResult(this, true);
     const queryStoreValue = this.queryManager.getApolloState().queries[this.queryId];
@@ -152,6 +159,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
       data,
       loading: isNetworkRequestInFlight(networkStatus),
       networkStatus,
+      partial,
     };
   }