+++ src/core/QueryManager.ts
@@ -83,8 +83,8 @@ import {
 } from '../data/readFromStore';
 
 import {
-  MutationBehavior,
   MutationQueryReducersMap,
+  MutationQueryReducer,
 } from '../data/mutationResults';
 
 import {
@@ -221,14 +221,12 @@ export class QueryManager {
   public mutate<T>({
     mutation,
     variables,
-    resultBehaviors = [],
     optimisticResponse,
-    updateQueries,
+    updateQueries: updateQueriesByName,
     refetchQueries = [],
   }: {
     mutation: DocumentNode,
     variables?: Object,
-    resultBehaviors?: MutationBehavior[],
     optimisticResponse?: Object,
     updateQueries?: MutationQueryReducersMap,
     refetchQueries?: string[],
@@ -247,15 +245,13 @@ export class QueryManager {
       operationName: getOperationName(mutation),
     } as Request;
 
-    // Right now the way `updateQueries` feature is implemented relies on using
-    // `resultBehaviors`, another feature that accomplishes the same goal but
-    // provides more verbose syntax.
-    // In the future we want to re-factor this part of code to avoid using
-    // `resultBehaviors` so we can remove `resultBehaviors` entirely.
-    const updateQueriesResultBehaviors = !optimisticResponse ? [] :
-      this.collectResultBehaviorsFromUpdateQueries(updateQueries, { data: optimisticResponse }, true);
-
     this.queryDocuments[mutationId] = mutation;
+
+    // Create a map of update queries by id to the query instead of by name.
+    const updateQueries: { [queryId: string]: MutationQueryReducer } = {};
+    Object.keys(updateQueriesByName || {}).forEach(queryName => (this.queryIdsByName[queryName] || []).forEach(queryId => {
+      updateQueries[queryId] = updateQueriesByName[queryName];
+    }));
 
     this.store.dispatch({
       type: 'APOLLO_MUTATION_INIT',
@@ -265,8 +261,8 @@ export class QueryManager {
       operationName: getOperationName(mutation),
       mutationId,
       optimisticResponse,
-      resultBehaviors: [...resultBehaviors, ...updateQueriesResultBehaviors],
       extraReducers: this.getExtraReducers(),
+      updateQueries,
     });
 
     return new Promise((resolve, reject) => {
@@ -285,11 +281,8 @@ export class QueryManager {
             document: mutation,
             operationName: getOperationName(mutation),
             variables,
-            resultBehaviors: [
-                ...resultBehaviors,
-                ...this.collectResultBehaviorsFromUpdateQueries(updateQueries, result),
-            ],
             extraReducers: this.getExtraReducers(),
+            updateQueries,
           });
 
           // If there was an error in our reducers, reject this promise!
@@ -839,52 +832,6 @@ export class QueryManager {
       variables: queryOptions.variables,
       document: transformedDoc,
     };
-  }
-
-  private collectResultBehaviorsFromUpdateQueries(
-    updateQueries: MutationQueryReducersMap,
-    mutationResult: Object,
-    isOptimistic = false,
-  ): MutationBehavior[] {
-    if (!updateQueries) {
-      return [];
-    }
-    const resultBehaviors: any[] = [];
-
-    Object.keys(updateQueries).forEach((queryName) => {
-      const reducer = updateQueries[queryName];
-      const queryIds = this.queryIdsByName[queryName];
-      if (!queryIds) {
-        // XXX should throw an error?
-        return;
-      }
-
-      queryIds.forEach((queryId) => {
-        const {
-          previousResult,
-          variables,
-          document,
-        } = this.getQueryWithPreviousResult(queryId, isOptimistic);
-
-        const newResult = tryFunctionOrLogError(() => reducer(
-          previousResult, {
-            mutationResult,
-            queryName,
-            queryVariables: variables,
-          }));
-
-        if (newResult) {
-          resultBehaviors.push({
-            type: 'QUERY_RESULT',
-            newResult,
-            variables,
-            document,
-          });
-        }
-      });
-    });
-
-    return resultBehaviors;
   }
 
   // Takes a set of WatchQueryOptions and transforms the query document