+++ test/store.ts
@@ -135,6 +135,8 @@ describe('createApolloStore', () => {
   });
 
   it('can reset itself and keep the observable query ids', () => {
+    const queryDocument = gql` query { abc }`;
+
     const initialState = {
       apollo: {
         data: {
@@ -155,6 +157,7 @@ describe('createApolloStore', () => {
           'networkError': null as any,
           'previousVariables': undefined as any,
           'queryString': '',
+          'document': queryDocument,
           'returnPartialData': false,
           'variables': {},
           'metadata': null,
@@ -174,7 +177,7 @@ describe('createApolloStore', () => {
       type: 'APOLLO_QUERY_INIT',
       queryId: 'test.0',
       queryString: '',
-      document: gql` query { abc }`,
+      document: queryDocument,
       variables: {},
       forceFetch: false,
       returnPartialData: false,
@@ -220,11 +223,7 @@ describe('createApolloStore', () => {
       mutationId: '1',
       optimisticResponse: {data: {incrementer: {counter: 1}}},
     });
-    const throwingBehavior: any = [
-      {
-        type: 'UnknownBehavior',
-      },
-    ];
+
     store.dispatch({
       type: 'APOLLO_MUTATION_RESULT',
       result: {data: {incrementer: {counter: 1}}},
@@ -232,10 +231,10 @@ describe('createApolloStore', () => {
       operationName: 'Increment',
       variables,
       mutationId: '1',
-      resultBehaviors: throwingBehavior,
+      extraReducers: [() => { throw new Error('test!!!'); }],
     });
 
-    assert(/UnknownBehavior/.test(store.getState().apollo.reducerError));
+    assert(/test!!!/.test(store.getState().apollo.reducerError));
 
     const resetState = {
       queries: {},
@@ -245,6 +244,16 @@ describe('createApolloStore', () => {
         {
           data: {},
           mutationId: '1',
+          action: {
+            type: 'APOLLO_MUTATION_RESULT',
+            result: {data: {data: {incrementer: {counter: 1}}}},
+            document: mutation,
+            operationName: 'Increment',
+            variables: {},
+            mutationId: '1',
+            extraReducers: undefined as undefined,
+            updateQueries: undefined as undefined,
+          },
         },
       ],
       reducerError: (null as Error | null),